// src/app/product/models/product.inventory.price.model.ts
export interface ProductInventoryPrice {
  DisplayOrder: number;
  EndDateTimeUtc: any;
  GeolocationId: number;
  GeolocationName: any;
  CurrencyId: number;
  CurrencyName: any;
  HTSDescription: string;
  HTSNumber: string;
  Id: number;
  IncotermDescription: any;
  IncotermId: number;
  IncotermMetaDescription: any;
  IncotermName: any;
  OldPrice: number;
  Price: number;
  ProductGuid: string;
  ProductInventoryId: number;
  ProductName: any;
  QuantityFrom: number;
  QuantityTo: number;
  SpecialAttributeValue: any;
  StartDateTimeUtc: any;
  UnitId: number;
  UnitName: string;
  WareHouseName: any;
}